+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [1.1.0-alpha.4]
+### CHANGED
+#### Menus
+- Changed the Title Screen
+- Changed the World Selection Screen layout
+
 ## [1.1.0-alpha.3] - 11/23/2025
 ### Added
 #### Blocks